% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreep.R
\name{clusterCells}
\alias{clusterCells}
\title{Cell Clustering}
\usage{
clusterCells(
  dreep.data,
  data,
  store.graph = T,
  seed = 180582,
  verbose = TRUE,
  resolution = 0.1,
  n.start = 50,
  n.iter = 250
)
}
\description{
Cell Clustering using cell-to-cell distances computed from the drug profiles
}
