% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreep.R
\name{runDiffDrugAnalisys}
\alias{runDiffDrugAnalisys}
\title{Differential Drug Analysis}
\usage{
runDiffDrugAnalisys(
  dreep.data,
  cell.set1,
  cell.set2,
  fdr.th = 0.1,
  show.plot = T
)
}
\description{
Identify a set of compounds specific for two cell populations
}
