% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreep.R
\name{runDREEP}
\alias{runDREEP}
\title{Drug Response Estimation from single-cell Expression Profiles (DREEP)}
\usage{
runDREEP(
  M,
  n.markers = 250,
  cores = 0,
  gsea = "multilevel",
  gpds.signatures = c("CTRP2", "GDSC")
)
}
\description{
Drug Response Estimation from single-cell Expression Profiles
}
