% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreep.R
\name{runDrugReduction}
\alias{runDrugReduction}
\title{Cell Dimensionality Reduction}
\usage{
runDrugReduction(
  dreep.data,
  fdr.th = 0.1,
  drug.subset = NULL,
  cores = 0,
  seed = 180582,
  verbose = T,
  reduction = "umap",
  cellDistAbsolute = T,
  storeCellDist = T,
  ...
)
}
\description{
Cell Dimensionality Reduction in the drug space
}
